add grapgql schema for post and get of book details

+++ src/main/java/com/example/demo/Repo/BookRepo.java
@@ -4,7 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.Model.Book;
 
-public interface BookRepo extends JpaRepository<Book, Long> {
+public interface BookRepo extends JpaRepository<Book, Integer> {
+
+    
     // Custom queries can be added here if needed
 }
